KSED/requirements.txt
pytest

allure-pytest

selenium

termcolor

requests

page-object

page-objects

urllib3

pytest-selenium